[BUGFIX] Use correct order of SQL parameters in SysRefindexHashUpdater
Supply the positional parameter arguments to the update statement in 
the correct order. The statement used for the updates reads
'UPDATE `sys_refindex` SET `hash` = ? WHERE `hash` = ?'
and thus the parameters should be in the order new hash, old hash, 
not the other way around.
It's the final order in the statement that's relevant when supplying the 
parameters, not the order in which they are defined in PHP.

Changed the definition order in PHP to match the order in the final 
SQL statement as well.

Change-Id: I7829bbc835a6b0f263b61fa07b2e39b36d846275
Resolves: #80766
Releases: master
Reviewed-on: https://review.typo3.org/52397
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Mona Muzaffar <mona.muzaffar@gmx.de>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>