[BUGFIX] Do not ask for current password if admin
Hide the field for the current password, if the current user is
admin anyway. If the admin changes the password for someone else
he/she doesn't know the current password of this person anyway.

Besides that, a password can always be changed without providing
a current password by an admin using the BE users module.

Resolves: #80951
Releases: master, 8.7, 7.6
Change-Id: Ic9673197e651df5d1fe5886c4b6f6f76a303062f
Reviewed-on: https://review.typo3.org/52556
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-on: https://review.typo3.org/52557