[BUGFIX] Linkvalidator: Scheduler task not working correctly
The scheduler task does not check anything but just returns and does not
send an email. Furthermore it is not possible to enter email address
separated by newlines. Under some conditions it can happen, that result
numbers are added to the last result.

Explode pid list into array in LinkAnalyzer::init and use the array later.
Check if email adresses are comma or newline separated when validating
them when saving the task. Add file link statistic to default email
template. Reset the link statistics before the run to avoid adding it to
the last result numbers.

Resolves: #81000
Releases: master, 8.7
Change-Id: I05a98c9babcb889a83ce2f1ec7c0a2bee5dafede
Reviewed-on: https://review.typo3.org/52593
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christer V <cvi@systime.dk>
Tested-by: Christer V <cvi@systime.dk>
Reviewed-by: Elmar Hinz <t3elmar@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>