[BUGFIX] Load PageTSconfig in localconf for LinkHandler
The pageTSconfig for LinkHandler is only loaded in ext_tables.php
which means it is only evaluated in the Frontend when a Backend User
is logged in.

However the functionality did work before, if a compiled pageTS was
available in the cache. This could have happened that the pageTS cache
was filled while a BE user was logged in, so it worked (but only in
random cases).

Resolves: #81057
Releases: master, 8.7, 7.6
Change-Id: Ib7cd62f82dff3a3b891dcb36e85c36b9a062142a
Reviewed-on: https://review.typo3.org/52648
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Marc Bastian Heinrichs <typo3@mbh-software.de>
Tested-by: Marc Bastian Heinrichs <typo3@mbh-software.de>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>