[BUGFIX] TcaMigration should not skip on existing overrideChildTca
If an overideChildTca key is already added by a TYPO3 core
function like ExtensionManagementUtility::getFileFieldTCAConfig()
the TCA migration should not be skipped. Otherwise the config will
be incomplete and result in missing fields.

The migration is changed that old "foreign_types",
"foreign_selector_fieldTcaOverride" and "foreign_record_defaults"
are merged into new "overrideChildTca", but the new settings
are keps (take precedence) if given.

Resolves: #80719
Resolves: #81045
Releases: master, 8.7
Change-Id: Ic8a0338e0a9882b3b1e7abcd3123fbd7aa2b3505
Reviewed-on: https://review.typo3.org/54582
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>