[BUGFIX] Keep MP parameters upon redirect of overlay
When a mount point, that is overlayed, redirects to the overlaying page,
don't throw away existing mount point parameters. The other mount point
parameters may specify a mount point upwards in the root line.

Under normal circumstances the page should not be linked at all. However
a redirect is provided and as it is provided it should work as expected.

Releases: master
Resolves: #81251
Change-Id: I4cf830e5b24ebc9799e58f83d7787a726f92a3d4
Reviewed-on: https://review.typo3.org/52829
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>