[BUGFIX] Accept configuration for external plugins from presets in ckeditor
Currently the default configuration of an external
plugin is always used for the configuration. This patch takes care
of a proper merging of the configuration.

The configuration of the extension is taken as default. If a
custom configuration in a preset is available, both configurations
will be merged.

Resolves: #81263
Releases: master,8.7
Change-Id: I4fc7c4f6f6ab40fe06e495e3ff98e067427c7d03
Reviewed-on: https://review.typo3.org/56157
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>