[BUGFIX] Avoid unnecessarily fetching file for processing
Files that are not configured as images, won't be processed,
but are still unnecessarily fetched for processing.

To avoid this, we change the processing API, that fetching
only happens, if the file is then used by following code.

In addition we avoid showing a non descriptive thumb
in FileInfoElement, which is used when editing meta data of the file.

Resolves: #81279
Releases: master, 8.7, 7.6
Change-Id: Ie68aa02b1b6149ddb7fbda346ac312d3646e260c
Reviewed-on: https://review.typo3.org/53033
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>