[TASK] Remove duplicate ViewHelper from extensionmanager
JSON encoding ViewHelper from Fluid is fully compatible.
Also removes a number of unused arguments in the Fluid
template - note that removed ViewHelper declared it was
a tag based ViewHelper which supports additionalAttributes,
but did not render a tag, which is why all the attributes were
simply ignored and unused.

Change-Id: I8d59f5a5e8971aac2b63be86c0a51aaf56437306
Resolves: #81301
Releases: master
Reviewed-on: https://review.typo3.org/52903
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>