[BUGFIX] Make encoding of parameters for cHash independent of PHP settings
http_build_query() internally uses the PHP setting arg-separator.output
to identify the URL parameter separator.
Our code relies on having a simple ampersand as this character.
If the PHP is set to a different configuration the code will fail.

This is fixed by enforcing the ampersand. Moreover we force RFC3986
encoding to ensure that decoding with rawurldecode() does the right
thing.

Change-Id: I92f9b877201c634303a2e387d9aad49106151f63
Releases: master, 8.7
Resolves: #81293
Reviewed-on: https://review.typo3.org/52923
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>