[BUGFIX] Remove disabled button in inline records
If the info icon is disabled it is now removed completely.
Previously the logic for new records was not caught correctly, so that
the case for the enabled icon with a new record was executed.
Disabling the icon should remove it completely,
as a disabled non-clickable icon is not helpful and confusing for editors.
This is now done, and the case for the enabled icon with new records is
also correctly caught.

Resolves: #81380
Releases: master, 8.7
Change-Id: I1c78a99a83efa3c88e3a6337df9cd0c8a959ea7b
Reviewed-on: https://review.typo3.org/53029
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>