[BUGFIX] Allow processing of multiple new record localizations
The current implementation of DataMapProcessor to determine the necessity
of synchronizing record localizations is too strict since it expects real
persisted database records and cannot resolve new records that have been
handed with the very same data-map to the DataHandler.

This mentioned constraint is resolved. Besides that, an additional value
processing for internal fields is added. This method resolves new record
ids to their real persisted UIDs using the famous remap stack. This only
is executed, if these fields have not been processed in a relation-aware
context and have a meaning to TYPO3 data-structures - for instance this
is the case for 'l10n_source' field defined as TCA type 'passthrough'.

Change-Id: Id1291910d85b3d314af7203314b8696a337fe364
Resolves: #80239
Releases: master, 8.7
Reviewed-on: https://review.typo3.org/52871
Reviewed-on: https://review.typo3.org/53211
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>