[BUGFIX] Extbase: Remove invalid class ObjectManagerException
There is a class called "ObjectManagerException" which was
introduced very early with Extbase and was a clone of
the (now further developed) object manager.

It was originally used in the wrong place in workspaces
and then changed in workspaces to the real object manager
but the class was kept.

This class was then wrongly renamed in TYPO3 6.0 with the
change to namespaces and called "Exception" which it is
definitively NOT and thus, plain wrong in the place as it
is.

The class (and the interface) should be removed as it
is not just plain useless and a hard mistake we carried on
for 7 years now.

Resolves: #52825
Releases: master, 8.7
Change-Id: I615caaa9d297631a338ca6488bfe8e5c29594e3b
Reviewed-on: https://review.typo3.org/53231
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>