[BUGFIX] Remove possibly undefined constant from test bootstraps
The constant LF might not be defined in functional and unit test
bootstrap by the time the script terminates with an error message.

Replacing the usage of this constant with the native value for a
linebreak causes the same result for the error message, but prevents
the additional and confusing messages about the not existing constant.

Change-Id: I66bd2092875e33fa8bb93a2a57c54daeaeb15ecf
Resolves: #81599
Releases: 7.6
Reviewed-on: https://review.typo3.org/53239
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>