[TASK] bamboo: Add a nightly build plan spec
Configures a second build for core master that
runs as nightly plan and has more combinations:
The acceptance and functionals are explicitly
executed on both php 7.0 and 7.1 with all DBMS,
the cgl checker runs through whole core.
This "full run" is too much for pre-merge tests,
so the nightly now takes care of these time
consuming things.
Reporting will be added if that plan is stable.
https://bamboo.typo3.com/browse/CORE-GTN

Change-Id: Id77fc0acceca7af67960b67fd8b8700c98df888e
Resolves: #81660
Releases: master, 8.7, 7.6
Reviewed-on: https://review.typo3.org/53304
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>