[!!!][TASK] Drop ext:func
The extension "func" which brings the "Web->Functions" main module
acts as a container for other extensions to hook in. Not too many
3rd party extensions used that over the years, though.
With the integration of wizard_sortpages and wizard_crpages into
ext:backend and the context menu on pages, ext:func is not used in
the core itself anymore.
The patch drops ext:func from core and adds an upgrade wizard to
fetch the extension from TER if still needed for 3rd party extensions
that did not get their dependencies right, yet.

Change-Id: If092953061a3b927ac93654a837522b5d32c2f1b
Resolves: #81787
Releases: master
Reviewed-on: https://review.typo3.org/53397
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>