[BUGFIX] Bring RTE linkbrowser options back
The CKEditor link browser handled the configuration array still as a
TypoScript array with dots. But the parsed YAML config is a plain array
now.

This patch makes sure the BrowserLinkController handles this config like
a normal array. Further it fixes the missing JavaScript part to set the 
link title based on selected css class.

Resolves: #81381
Releases: master, 8.7
Change-Id: Ie4749f820082e185208bf811aad7133bfc82d37a
Reviewed-on: https://review.typo3.org/53481
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Reviewed-by: Markus Sommer <markus.sommer@typo3.org>
Tested-by: Markus Sommer <markus.sommer@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>