[BUGFIX] Clone ObjectStorage in ObjectAccess
The ObjectStorage Object passed to iterator_to_array() ends up
with the pointer at the last item, so that any ongoing iteration
through the ObjectStorage wrapping the call of ObjectAccess is
interrupted.
This patch ensures that the ObjectStorage is cloned before converted
to an array so that the original object keeps it original iteration
state.

Resolves: #81597
Relates: #66995
Releases: master, 8.7
Change-Id: Ied025ff616e100cc5eb5dedd5b7b6a95293ddfcf
Reviewed-on: https://review.typo3.org/53525
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Claus Due <claus@phpmind.net>
Reviewed-by: Sascha Rademacher <sascha.rademacher@gmail.com>
Tested-by: Sascha Rademacher <sascha.rademacher@gmail.com>
Reviewed-by: Benjamin Kluge <b.kluge@neusta.de>
Tested-by: Benjamin Kluge <b.kluge@neusta.de>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>