[FEATURE] Store extension configuration as plain array
Extension configuration is now stored as plain array
instead of serialized values. To ensure backwards-
compatibility and stream-line core usage, the old
values will still be stored and written in
$GLOBALS['TYPO3_CONF_VARS']['EXT']['extConf'] whereas
the new array will be stored in
$GLOBALS['TYPO3_CONF_VARS']['EXTCONF'].

As a second step we are going to introduce an API for
retrieving extension configuration to remove the necessity
for GLOBALS access in this case.

Resolves: #82254
Releases: master
Change-Id: I52ef7768491633e114e7e1b153a4ba63e07243ff
Reviewed-on: https://review.typo3.org/53657
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Joerg Kummer <typo3@enobe.de>
Tested-by: Joerg Kummer <typo3@enobe.de>
Reviewed-by: Romain Canon <romain.hydrocanon@gmail.com>
Tested-by: Romain Canon <romain.hydrocanon@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>