[TASK] Move install tool entry point to typo3/install.php
To streamline and unify the entry scripts, the install tool
entry point is now moved to typo3/install.php

Calling the old entry scripts and the shortcut typo3/install/
is deprecated.

All access to public resources is changed to properly use
the resource view helper, so that the path to the assets
are not hard coded any more.

Resolves: #81951
Releases: master
Change-Id: Ia8827130bb041802e743767907d16269cda2fef0
Reviewed-on: https://review.typo3.org/53579
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>