[BUGFIX] CSS Styled Content should render image captions
When adding two CEs with one image each (textpic), the
TypoScript register is using faulty data from file:current:description
for the caption.

This behaviour was fixed with #77645 but now shows that
CSS Styled Content did use this part in a wrong way instead
of using the custom register that is explictly used before.

Resolves: #81866
Releases: 8.7, 7.6
Change-Id: Ia61556a7e65d6a5b2ed0e13db0fbd876c279356e
Reviewed-on: https://review.typo3.org/53587
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>