[BUGFIX] YouTubeRenderer "controls" option
The "controls" parameter appended to YouTube embed URL is either set to
value provided in $options array or to default backwards-compatible
value (controls=2).

Specifically, if controls option is set to 0, the controls parameter
must be present in URL, because default value for it is 1.

Fix is applied to related unit test too.

Resolves: #82044
Releases: master, 8.7, 7.6
Change-Id: I1519b5f515f85eb473f590762171b250d26f32c4
Reviewed-on: https://review.typo3.org/53650
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>