[BUGFIX] Remove symlink to typo3 folder in web dir
When upgrading from older TYPO3 versions, composer installation
from that version created a symlink to the typo3 folder.

Before trying to remove typo3/sysext, we need to ensure
to remove a potentially existing symlink to typo3 folder
to not accidentally delete that sysext folder within
the sources directory.

Resolves: #82090
Releases: master
Change-Id: I12ef27d6236a9dbe844803afce55247a23ee5358
Reviewed-on: https://review.typo3.org/53682
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Tested-by: Mathias Brodala <mbrodala@pagemachine.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>