[BUGFIX] Use DriverInterface instead AbstractDriver for method signature
FAL drivers are not obligated to extend the AbstractDriver but only
to implement the DriverInterface. Having custom drivers that are not
based on AbstractDriver (because they are f.e. non hierarchical)
currently causes an Exception because PreviewProcessing expects an
instance of AbstractDriver as method argument.

This patch changes te signature from AbstractDriver to DriverInterface.

Resolves: #82114
Releases: master, 8.7, 7.6
Change-Id: Ia27c6c1c7510ecfe5afd0f7d8bedf81bf3d31c84
Reviewed-on: https://review.typo3.org/53684
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>