[BUGFIX] _LOCAL_LANG overwrite via TypoScript with nested keys
If language labels are set for the top level key and sub
keys in TypoScript the top level key was not taken into
account as it got stored with 'typoScriptNodeValue' as key.

Special handling for that specific key was introduced, to
make overwriting of top level keys possible again.

Change-Id: Ibd5f5a939b1de530ba93840cfb7ba5ab5bc85c2f
Resolves: #81095
Releases: master, 8.7
Reviewed-on: https://review.typo3.org/53704
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>