[TASK] Install Tool: Split up Checks into RequestHandlers
In order to get the controllers free of security checks, the logic of
authentication/session handling is moved from various controllers into
the RequestHandler and the Application.

Additionally, a second RequestHandler (RecoveryRequestHandler) is introduced
which acts as a fallback if TYPO3 is not installed yet, or the installation is
broken (e.g. missing PackageStates.php).

This brings a cleaner dispatching mechanism, having the RecoveryRequestHandler
(which can handle any request) dealing with the StepController, and the
regular RequestHandler (with higher priority) for handling the maintenance
functionality for running installations.

Resolves: #82238
Releases: master
Change-Id: I906ab3f15479bbe70202608f6bd7838cc22d9e3a
Reviewed-on: https://review.typo3.org/53800
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>