[TASK] Install Tool: Remove authentication from backend context
Currently calling the install tool modules from within the Backend does a
simple redirect with adding GET variables.

That's the reason why you need to re-authenticate again, and the context
is handed over as a query parameter, which is simply not needed at all.

Now, the redirect is removed, as the Backend entrypoint / request handler
handles the authentication of the backend user, and the standalone entry
point deals with the install tool password etc.

The context parameter is now detected by the entry point (!) as well,
allowing to get rid of quite some code.

There are some more consequences:
- Calling the install tool from the backend does not validate if you configuration
is set up (= recovery necessary) -> since you're already in the backend we guess
you're fine anyway.
- Redirect functionality is almost not needed anymore in the regular request handler
- routeParameters concept was removed again (which was introduced a couple of weeks ago)

Additionally, the contextService could be replaced at a later stage with just
a string.

Resolves: #82306
Releases: master
Change-Id: If7e4ddfaccf46cf93448d06c0ba9af81d5b9494c
Reviewed-on: https://review.typo3.org/53860
Reviewed-by: Romain Canon <romain.hydrocanon@gmail.com>
Tested-by: Romain Canon <romain.hydrocanon@gmail.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>