[BUGFIX] Check if LocalConfiguration.php is writable
In TYPO3 v8, the ConfigurationManager only checks if LocalConfiguration.php
is writaeble and not if typo3conf itself is writable.

Optimizing this check removes a warning in an environment where typo3conf/
is hardened to only have LocalConfiguration.php be modified and nothing else.
This change helps running TYPO3 v7 on platform.sh with fewer error messages.

You can compare this line of code with the same line in TYPO3 v8/master.

Resolves: #82289
Releases: 7.6
Change-Id: I7adca89eaa0d9f5b28e3f034f9452d97591a0d80
Reviewed-on: https://review.typo3.org/53878
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>