[BUGFIX] Unify Extbase base command error
This changes the output of "typo3/sysext/core/bin/typo3 extbase"
to show the standard Symfony error message instead of
a misleading one.

Calling a Symfony base command outputs an error message
with a list of possible commands.

We change the command name of the core command to
get the same behavior for the Extbase base command.

Resolves: #81917
Releases: master, 8.7
Change-Id: I9ba05600a461aceb4180d3e2dd124a5ec3006e08
Reviewed-on: https://review.typo3.org/54162
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>