[TASK] Use LanguageService inside TSFE
TSFE does the same logic as $GLOBALS['LANG'] (= LanguageService) for
fetching labels and language files, but with a different language key.

However, LanguageService is not a singleton, so it _can_ be used
multiple times at the same request, so why not use the LanguageService
for labels based on config.sys_language, and the $GLOBALS['LANG'] record
based on the current backend user.

So, TSFE only acts as a wrapper for LanguageService instead of having
duplicate code.

A few now useless internal properties are dropped, some other
methods are deprecated:

- TSFE->LL_files_cache
- TSFE->LL_labels_cache

- TSFE->readLLfile()
- TSFE->getLLL()
- TSFE->initLLvars()

Resolves: #81217
Releases: master
Change-Id: Iae85ac2eda4d7f77960de2d66b9069f6017c31d8
Reviewed-on: https://review.typo3.org/52799
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>