[BUGFIX] Install tool: Use standalone application in backend context
With recent routing changes the install tool "backend context" has
been switched to a "normal" backend module loaded directly from
within the backend application. The standalone is an own application
with its own bootstrap.
This leads to various issues in the install tool since the backend
and standalone requests lead to different bootstrap states, for
instance TCA is initialized in BE context, but isn't in standalone.
Within backend, this won't change until the backend application can
bootstrap to different states depending on a controller.
To solve install tool related issues for now, the BackendModuleController
of the install tool called from within backend now starts a casual
install tool session and marks this session as "initiated from a
valid system maintainer". It then redirects to the standalone
application which omits login and enable install tool file check if
the session is marked as backend user session.

Change-Id: I352e6d04e7a91c56ccf2383f784ae94464c9aacd
Resolves: #82448
Related: #82306
Releases: master
Reviewed-on: https://review.typo3.org/54111
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>