[BUGFIX] Fix link rendering in li tag
Changes the rendering for li tags in RTE in sys ext css_styled_content.
This change is based on a former bugifx for TYPO3 6.2:
https://review.typo3.org/#/c/38681/

The fix was not ported to css_styled_content for TYPO3 7,
so the issue is present in TYPO3 7 if no fluid_styled_content
TypoScript is present.

Releases: 7.6
Resolves: #65506
Change-Id: I1aa744b7f381681a01cb797ff6dd9c61f221934a
Reviewed-on: https://review.typo3.org/54165
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>