[TASK] Raise default jpg image quality to 85
DefaultConfiguration setting is 70 (default since 12 years), but the
FactoryConfiguration setting is set to "80" for new installations.

This change raises all image quality settings to "85" if not overridden
by any installation (= new installation in the last 10 years or not changed
via the install tool), as we don't have a big file size issue with JPGs anymore
and we don't have duplicates anymore
(due to FAL instead of uploads/pics/mypic_13.jpg) and the internet connections
got better.

Resolves: #82655
Releases: master
Change-Id: Ibcb38ece4038a414bfcae103463c727ed3cf65c0
Reviewed-on: https://review.typo3.org/54267
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>