[!!!][TASK] Remove option to generate PNGs limited to 64 colors
The global configuration "png_truecolor" which ensured that generated PNG
images only have 64 colors (instead of the max colors) came from the transition
12 years ago to keep a small file size for PNG images.

The option was turned on for new installations and existing installations,
and if not explicitly disabled, generated true-color PNG images at any time.

The option did not have any effect for FAL-related image resizing at all.

In order to generate PNGs with a limited color-space, it is still possible to do so with
additional "reduceColors" option, as well as adding additional parameters to
ImageMagick when needed.

However, having a global option for this feature seems to be not practical these
days anymore.

Resolves: #82680
Releases: master
Change-Id: I3e00cc1c37a5ed8287dcf7cad49d2a2ab10730e6
Reviewed-on: https://review.typo3.org/54277
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>