[BUGFIX] Delete unavailable "recently switched to" users from UC
If a user who appears in the "Recently switched to" list became
unavailable for some reason, only the placeholder avatar was rendered,
giving a slight hint of life in the past.

Such users are now deleted from UC and do not appear in the list anymore.
Furthermore, the section label was slightly modified to be more
understandable.

For performance reasons, only one query is triggered now and sorting is
done based on the array of UIDs stored in UC.

Resolves: #82796
Related: #80581
Releases: master
Change-Id: Ieff8a779639f277de1c0cb7feaa0fe2b82bd97c5
Reviewed-on: https://review.typo3.org/54433
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Joerg Boesche <typo3@joergboesche.de>
Tested-by: Joerg Boesche <typo3@joergboesche.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>