[BUGFIX] Ensure correct page is selected in page tree
There are several actions in the backend that should
trigger a certain page to be selected in the tree.

Among others these are: open documents toolbar, live search,
favorites.

Fix every scenario in which these actions are triggered
(e.g. page tree was already initialized or not), make
sure the desired page will always be selected and
properly store the selection in the various states
so that the selection persists tree refreshes.

In addition we ensure that the content frame will
be called with the correct id, if the module uses
the page tree as navigation component.

Releases: master, 8.7
Resolves: #82877
Change-Id: I1904c02b2246f01202c6607a9b718aebe0e33341
Reviewed-on: https://review.typo3.org/54491
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Tested-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>