[BUGFIX] FormEngine: returnUrl handling in inline
If an inline child, opened via ajax request, generates links leaving
the current records and opening a different one, on closing that
record the returnUrl is wrong and points to the url of the inline
ajax request.
The original return url of the main record is lost in the compile chain.
The returnUrl data provider sets $data['returnUrl'] to current url if not
otherwise given.
The patch hands over the original return url to the inline ajax controller
and lets it 'sink down' in the data compile chain to inline children so
the original return url is available when container or elements generate urls.
Additionally some broken 'returnURL' handling in InlineRecordContainer
and jsfunc.inline.js is cleaned up and removed.

Change-Id: Ia47dc8cee068fc9c4d1c5e16c1960b1a58fcb2f5
Resolves: #82525
Releases: master, 8.7
Reviewed-on: https://review.typo3.org/54543
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Kjeld Schumacher <k.schumacher@familie-redlich.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>