[TASK] Do not silently ignore exceptions in ImageViewhelper
Exceptions in the ImageViewhelper are currently silently ignored.
So you'll get no information if a wrong value has been passed
to the src attribute or if there went something wrong.
The exception is not rethrown as Fluid viewhelper exception so
it will be catched and logged in Production Context by the
AbstractViewhelper that calls the render method.

Change-Id: I95124ebb945366d260b4351ccb85d61a399c50b5
Resolves: #82918
Releases: master
Reviewed-on: https://review.typo3.org/54561
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Manuel Selbach <manuel_selbach@yahoo.de>
Reviewed-by: Jan Stockfisch <jan.stockfisch@googlemail.com>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>