[BUGFIX] Keep state of extendedSearch
formAction has set a bunch of variables that are required by
the frontend for the extendedSearch

Inside searchAction this parameters were not set resulting in
the extendedSearch section being empty in frontend after searching.

With this change the processing of the variables is extracted
into a method and called from formAction and searchAction.

Releases: master, 8.7
Resolves: #82064
Change-Id: Ia21c445ef0aa72e8e084e9bd021bc5caa7a1e74b
Reviewed-on: https://review.typo3.org/54646
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>