[BUGFIX] Set l10n_state before synchronizing fields
L10nModeUpdater checks whether fields in translated records are
modified, thus not synchronized with the parent record and the
new l10n_state for this field is set to "custom" in this case
or "parent" if the field is empty.

However this is done after the data is synchronized, which
leads to wrong data being written by the DataMapProcessor.

To fix this, we move setting l10n_state upwards in the code,
so that relations and fields from translated records can be
updated correctly.

Resolves: #83006
Releases: master, 8.7
Change-Id: Ibb06eecb6cf743bd6bba1990ca6f9391c983821f
Reviewed-on: https://review.typo3.org/54656
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>