[BUGFIX] IRRE actions should render error messages from DataHandler
When IRRE actions (localize/synchronize) are triggered that result in a
DataHandler error (e.g. "Localization failed, "languageField" and
"transOrigPointerField" must be defined for the table") because of a missing
configuration, the current implementation does not evaluate DataHandler
results in the JSON response.

The change adds the same functionality as for other DataHandler calls
to evaluate the JSON response.

Resolves: #83011
Releases: master, 8.7
Change-Id: I7a2069b78cd9c3fa031b7ac194ae5b18596b6a3e
Reviewed-on: https://review.typo3.org/54659
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>