[BUGFIX] Only update fe_users.is_online if user is logged in
The field is_online is updated in TSFE right after a user has authenticated,
but if the user has no usergroup, he/she is not logged in.

The update should be moved in the TSFE->fe_user and the check
should only be called when a logged-in user is found.

Resolves: #83086
Releases: master
Change-Id: I14b31662d99772b29e0e37153d44925b4309d69c
Reviewed-on: https://review.typo3.org/54756
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Henning Liebe <h.liebe@neusta.de>
Reviewed-by: Joerg Boesche <typo3@joergboesche.de>
Reviewed-by: Łukasz Uznański <l.uznanski@macopedia.pl>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>