[BUGFIX] UpgradeWizard check for utf8 is wrong
The upgrade wizard checking for utf-8 sets a
variable called charsetOk which worked
differently on MySQL compared to other DBMS.

On MySQL it contained true if the charset was
not OK, in all other cases it was true if ok.

This is now consistently handled according to
the variable name.

Change-Id: Iee347b6c2469ad9a33e9bd4109d361dc337a63a0
Resolves: #83141
Releases: master
Reviewed-on: https://review.typo3.org/54835
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>