[BUGFIX] Hint for 'Check for broken extensions' if install tool crashes
If the install tool crashes in cards that load ext_* files from extensions,
it currently clears html body and only renders a message "Something went wrong".

This will later become a recovery view which can analyze where and how the
error happened to show the user good option on how to resurrect the instance.

For now, the message is extended a bit linking to the "Upgrade" menu saying
one should use the "Check for broken extensions". This way - until a bigger
solution evolved - we at least have a link to click on, and the user does not
need to fiddle with the url bar of the browser. This is especially useful
in install tool standalone mode where the whole content including the left
main menu is currently substituted by the error message.

Change-Id: I8e1c450949d9f00f8c2e0e1e95a847979973c124
Resolves: #82922
Releases: master
Reviewed-on: https://review.typo3.org/54559
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>