[BUGFIX] Undefined constant OCI_COMMIT_ON_SUCCESS
Test suite on core v7 with PHP 7.2 complains a lot about
"Undefined constant OCI_COMMIT_ON_SUCCESS". This is
found in the adodb oracle oci8 driver, used as property
default value and thus evaluated as soon as the driver class
is instantiated in tests.
The test suite does not include oci8 in php. The php module
sets constant OCI_COMMIT_ON_SUCCESS to hex '0x20' which is
32 in decimal.
The patch uses decimal 32 to suppress the warning if
oci8 module is not loaded for tests.

Change-Id: I218e9758fe8b1bf8a92a8398f6782faa4ed4c9c9
Resolves: #83209
Releases: 7.6
Reviewed-on: https://review.typo3.org/54923
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>