[FEATURE] Introduce conditional variants for form elements
Variants allow you to change properties of a form element.

Variants can contain conditions. If a concrete condition
is TRUE the variant is applied. If no condition exists the
variant will be ignored.

Variants make it possible to manipulate form element values,
validator options, and finisher options based on conditions.
This allows you among other things:

* translation of form element values depending on the frontend
  language
* set and remove validators of one form element depending on the
  value of another form element
* hide entire pages depending on the value of a form element
* set finisher values depending on the value of a form element
* hiding a form element in certain finishers and on the
  summary page

This feature implements variants for the frontend rendering and
the ability to define variants in the formDefinition. The
implementation to define variants in the form editor is out of
scope of this patchset.

Releases: master
Resolves: #84133
Change-Id: I9efeeea5af67df2d2f9252339c26baf8a03cf9c8
Reviewed-on: https://review.typo3.org/54982
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Björn Jacob <bjoern.jacob@tritum.de>
Tested-by: Björn Jacob <bjoern.jacob@tritum.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>