[!!!][TASK] Implement Salted Passwords against SaltInterface
Salt classes must implement SaltInterface only. The AbstractSalt
class is renamed to AbstractComposedSalt and implements
SaltInterface. Methods for salt-classes that compose the
password-hash string themselves (which are currently all in
saltedpasswords) are moved to AbstractComposedSalt as well.

This cleanup change prepares for the integration of the
PHP password API in a following change.

Relates: #79795
Relates: #79889
Resolves: #83294
Releases: master
Change-Id: Ife24aa39be99c5ad391b0f10497a2bceb04084f3
Reviewed-on: https://review.typo3.org/52737
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>