[BUGFIX] Repopulate localconfig before cache reload in ext mgr
Reloading caches requires the ext_localconf files. If in an
ext_localconf the extension configuration is accessed this
will fail directly after installation as the configuration is
not reloaded yet.

The patch changes the order to first write the new configuration,
then read the new configuration into the global array and
finally reloading the caches.

Change-Id: I9e2cd70078b2038aa30f2b4706aec46243ab2a9d
Resolves: #83300
Related: #82368
Releases: master
Reviewed-on: https://review.typo3.org/55045
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>